Fixed the file size update issue on file save
There was a bug where the file size display on status bar was not
updated when a file was saved. The users had to re-open the document in
order to update the file size display.

This patch fixes the issue by forcing status bar right text field to
update when a file size is updated. The performance impact of this
patch should be minimal.