Stop using showsControlCharacters flag
Since NSLayoutManager's showsControlCharacters flag is totally buggy (at leaset on El Capitan), we stopped using this since CotEditor 2.3.3 that will be released in 2016-01.
Previously, CotEditor used this flag for "Other Invisible Characters."
However as CotEditor draws such control-glyph-alternative-characters by itself in `drawGlyphsForGlyphRange:atPoint:`, this flag is actually not so necessary as I thougth. may be.