Add hook mechanics for AppleScript plugins
Script bundles can have `Info.plist`, which is useful to store metadata.
For example, the following key-value set can be used to determine
whether the event of document opened should delivered to the plugin:

```
<key>CotEditorHandlers</key>
<array>
<string>document opened</string>
</array>
```

By this change, the `CotEditorHandlers` in `Info.plist` of script bundle
lugins will be read and registered to `ScriptManager`.

Additionally, the API to invoke the registered handlers was introduced:
  `ScriptManager.dispatch(_:toHandlersAt)`