Add persistent mode for AppleScript scripts
`NSUserAppleScriptTask` is the recommended way to execute user-supplied
scripts and the only way to execute them outside of the sandbox.
However, `NSUserAppleScriptTask` has a performance issue because it
compiles a script and spawns a separate process every time the script
would be executed.

This change adds the persistent execution mode with `OSAScript`, which
compiles scripts only once but executes them inside the sandbox.