Restrict the refresh rate of the script menu
When many items under the application scripts directory are renewed
simultaneously, `presentedSubitemDidChange(at:)` is called for each
renewed file.
This will cause refreshing the script menu at too high rate when renewal
of many files are occurred while the application is active.

By this change, the refresh rate was restricted and high CPU load was
suppressed.